Move tests as well.
